fix(dd): cast file size to size_t for 32-bit /WX build
am_buf.resize(am_file.size()) narrowed a 64-bit file size to the 32-bit
size_t on the Win32 build, tripping C4267 which /WX promotes to an error
and broke the windows-x86 release leg at HEAD. Explicit static_cast on
both the resize and read_at length.

+++ src/engine/data_dict.cpp
@@ -207,8 +207,9 @@ util::Result<void> DataDict::load_add_binary_(const std::string& buf) {
                 auto am_file = std::move(amf_res).value();
                 auto sz_res = am_file.size();
                 if (sz_res && sz_res.value() > 0) {
-                    am_buf.resize(sz_res.value());
-                    auto rr = am_file.read_at(0, am_buf.data(), sz_res.value());
+                    am_buf.resize(static_cast<std::size_t>(sz_res.value()));
+                    auto rr = am_file.read_at(0, am_buf.data(),
+                                              static_cast<std::size_t>(sz_res.value()));
                     if (rr && rr.value() < am_buf.size()) am_buf.resize(rr.value());
                 }
             }
